requirements.txt
flask
flask-cors
pypdf
torch
transformers
google-genai
langchain
langchain-community
langchain-google-genai